requirements.txt
requests
beautifulsoup4
ddddocr
pycryptodome
